pyproject.toml
[project]
name = "dio-bank-fastapi"
version = "0.1.0"
description = ""
authors = [
    {name = "Luiz Eduardo",email = "luiz.rezendemoraes@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "pyjwt",
    "uvicorn[standard]",
    "databases[aiosqlite,asyncpg]",
    "psycopg2-binary",
    "pydantic-settings",
    "alembic"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
